tweak: Make this branch build.

+++ Code/components/console/StringTokenizer.h
@@ -22,7 +22,7 @@ class StringTokenizer
   private:
     std::string m_delim;
     std::string m_string;
-    int m_count;
+    size_t m_count;
     std::string::size_type m_begin;
     std::string::size_type m_end;
 };
